fix(frontend/controller): improve match highlighting + split up the search component (#594)
Fixes #567
Fixes #569
